fix warning

+++ contracts/deployment/Migrations.sol
@@ -4,20 +4,14 @@ pragma solidity ^0.8.19;
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 contract Migrations is Ownable {
-    uint public last_completed_migration;
-    mapping(uint => address) public migrationContracts;
+    uint256 public last_completed_migration;
+    mapping(uint256 => address) public migrationContracts;
     mapping(address => bool) public isMigrationContract;
 
     // Events
-    event MigrationCompleted(
-        uint indexed migrationNumber,
-        address indexed contractAddress,
-        uint timestamp
-    );
+    event MigrationCompleted(uint256 indexed migrationNumber, address indexed contractAddress, uint256 timestamp);
     event MigrationContractRegistered(
-        address indexed contractAddress,
-        uint indexed migrationNumber,
-        uint timestamp
+        address indexed contractAddress, uint256 indexed migrationNumber, uint256 timestamp
     );
 
     constructor() Ownable(msg.sender) {
@@ -29,52 +23,38 @@ contract Migrations is Ownable {
         _;
     }
 
-    function setCompleted(uint completed) public restricted {
-        require(
-            completed > last_completed_migration,
-            "Migration number must be greater than last completed"
-        );
+    function setCompleted(uint256 completed) public restricted {
+        require(completed > last_completed_migration, "Migration number must be greater than last completed");
         last_completed_migration = completed;
         emit MigrationCompleted(completed, address(0), block.timestamp);
     }
 
     function upgrade(address new_address) public restricted {
         require(new_address != address(0), "Invalid contract address");
-        require(
-            !isMigrationContract[new_address],
-            "Contract already registered as migration"
-        );
+        require(!isMigrationContract[new_address], "Contract already registered as migration");
 
         Migrations upgraded = Migrations(new_address);
         upgraded.setCompleted(last_completed_migration);
 
         // Register the new migration contract
-        uint nextMigration = last_completed_migration + 1;
+        uint256 nextMigration = last_completed_migration + 1;
         migrationContracts[nextMigration] = new_address;
         isMigrationContract[new_address] = true;
 
-        emit MigrationContractRegistered(
-            new_address,
-            nextMigration,
-            block.timestamp
-        );
+        emit MigrationContractRegistered(new_address, nextMigration, block.timestamp);
     }
 
     /**
      * @dev Get migration contract by number
      */
-    function getMigrationContract(
-        uint migrationNumber
-    ) external view returns (address) {
+    function getMigrationContract(uint256 migrationNumber) external view returns (address) {
         return migrationContracts[migrationNumber];
     }
 
     /**
      * @dev Check if address is a migration contract
      */
-    function isMigrationContractAddress(
-        address contractAddress
-    ) external view returns (bool) {
+    function isMigrationContractAddress(address contractAddress) external view returns (bool) {
         return isMigrationContract[contractAddress];
     }
 
@@ -84,16 +64,8 @@ contract Migrations is Ownable {
     function getMigrationStats()
         external
         view
-        returns (
-            uint lastCompleted,
-            uint totalMigrations,
-            bool hasActiveMigration
-        )
+        returns (uint256 lastCompleted, uint256 totalMigrations, bool hasActiveMigration)
     {
-        return (
-            last_completed_migration,
-            last_completed_migration,
-            last_completed_migration > 0
-        );
+        return (last_completed_migration, last_completed_migration, last_completed_migration > 0);
     }
 }
