+++ contracts/deployment/factories/GameFactory.sol
@@ -29,16 +29,12 @@ contract GameFactory is Initializable, OwnableUpgradeable {
         address gameAddress;
         GameType gameType;
         bool isActive;
-        uint createdAt;
+        uint256 createdAt;
     }
 
     // Events
-    event GameCreated(
-        address indexed gameAddress,
-        GameType gameType,
-        uint timestamp
-    );
-    event GameDeactivated(address indexed gameAddress, uint timestamp);
+    event GameCreated(address indexed gameAddress, GameType gameType, uint256 timestamp);
+    event GameDeactivated(address indexed gameAddress, uint256 timestamp);
 
     // State variables
     mapping(address => GameInfo) public games;
@@ -91,18 +87,13 @@ contract GameFactory is Initializable, OwnableUpgradeable {
         contractAddresses[1] = referralContract;
         contractAddresses[2] = statsContract;
 
-        registryInstance.registerBatchContracts(
-            contractNames,
-            contractAddresses
-        );
+        registryInstance.registerBatchContracts(contractNames, contractAddresses);
     }
 
     /**
      * @dev Create a new game contract
      */
-    function createGame(
-        GameType gameType
-    ) external onlyOwner returns (address) {
+    function createGame(GameType gameType) external onlyOwner returns (address) {
         address gameAddress;
 
         if (gameType == GameType.ONE_DAY) {
@@ -114,12 +105,8 @@ contract GameFactory is Initializable, OwnableUpgradeable {
         }
 
         // Register the game
-        games[gameAddress] = GameInfo({
-            gameAddress: gameAddress,
-            gameType: gameType,
-            isActive: true,
-            createdAt: block.timestamp
-        });
+        games[gameAddress] =
+            GameInfo({gameAddress: gameAddress, gameType: gameType, isActive: true, createdAt: block.timestamp});
 
         allGames.push(gameAddress);
 
@@ -163,10 +150,7 @@ contract GameFactory is Initializable, OwnableUpgradeable {
             address(registry) // registry address
         );
 
-        ERC1967Proxy proxy = new ERC1967Proxy(
-            sevenDaysImplementation,
-            initData
-        );
+        ERC1967Proxy proxy = new ERC1967Proxy(sevenDaysImplementation, initData);
 
         return address(proxy);
     }
@@ -184,16 +168,16 @@ contract GameFactory is Initializable, OwnableUpgradeable {
      * @dev Get all active games
      */
     function getActiveGames() external view returns (address[] memory) {
-        uint activeCount = 0;
-        for (uint i = 0; i < allGames.length; i++) {
+        uint256 activeCount = 0;
+        for (uint256 i = 0; i < allGames.length; i++) {
             if (games[allGames[i]].isActive) {
                 activeCount++;
             }
         }
 
         address[] memory activeGames = new address[](activeCount);
-        uint index = 0;
-        for (uint i = 0; i < allGames.length; i++) {
+        uint256 index = 0;
+        for (uint256 i = 0; i < allGames.length; i++) {
             if (games[allGames[i]].isActive) {
                 activeGames[index] = allGames[i];
                 index++;
@@ -206,19 +190,14 @@ contract GameFactory is Initializable, OwnableUpgradeable {
     /**
      * @dev Get game info
      */
-    function getGameInfo(
-        address gameAddress
-    ) external view returns (GameInfo memory) {
+    function getGameInfo(address gameAddress) external view returns (GameInfo memory) {
         return games[gameAddress];
     }
 
     /**
      * @dev Set implementation addresses
      */
-    function setImplementations(
-        address _oneDayImplementation,
-        address _sevenDaysImplementation
-    ) external onlyOwner {
+    function setImplementations(address _oneDayImplementation, address _sevenDaysImplementation) external onlyOwner {
         oneDayImplementation = _oneDayImplementation;
         sevenDaysImplementation = _sevenDaysImplementation;
     }
@@ -227,25 +206,23 @@ contract GameFactory is Initializable, OwnableUpgradeable {
         address[] storage allGamesArr = allGames;
         uint256 length = allGamesArr.length;
         address[] memory result = new address[](length);
-        for (uint i = 0; i < length; i++) {
+        for (uint256 i = 0; i < length; i++) {
             result[i] = allGamesArr[i];
         }
         return result;
     }
 
-    function getGameCount() public view returns (uint) {
+    function getGameCount() public view returns (uint256) {
         return allGames.length;
     }
 
-    function getGamesByType(
-        string memory /* gameType */
-    ) public view returns (address[] memory) {
+    function getGamesByType(string memory /* gameType */ ) public view returns (address[] memory) {
         address[] storage allGamesArr = allGames;
         uint256 length = allGamesArr.length;
         address[] memory temp = new address[](length);
         uint256 count = 0;
 
-        for (uint i = 0; i < length; i++) {
+        for (uint256 i = 0; i < length; i++) {
             // 여기서는 간단한 예시로 모든 게임을 반환
             // 실제로는 게임 타입을 체크하는 로직이 필요
             temp[count] = allGamesArr[i];
@@ -253,7 +230,7 @@ contract GameFactory is Initializable, OwnableUpgradeable {
         }
 
         address[] memory result = new address[](count);
-        for (uint i = 0; i < count; i++) {
+        for (uint256 i = 0; i < count; i++) {
             result[i] = temp[i];
         }
